Feature/tao 4412 refactor renderer

+++ model/BookletClassService.php
@@ -135,8 +135,8 @@ class BookletClassService extends tao_models_classes_ClassService
     {
         $property = $this->getProperty(self::PROPERTY_FILE_CONTENT);
         $contentUri = $instance->getOnePropertyValue($property);
-
-        if ($contentUri) {
+        
+        if ($contentUri && !($contentUri instanceof \core_kernel_classes_Literal)) {
             $storageService = $this->getServiceLocator()->get(StorageService::SERVICE_ID);
             $storageService->deleteFile($contentUri);
         }
